+++ src/Products/MeetingCharleroi/adapters.py
@@ -486,7 +486,7 @@ class CustomCharleroiMeetingItem(CustomMeetingItem):
             # item in state giveable but item not complete
             if item.query_state() in FINANCE_WAITING_ADVICES_STATES:
                 return {'displayDefaultComplementaryMessage': False,
-                        'displayAdviceReviewState': True,
+                        'displayAdviceReviewState': False,
                         'customAdviceMessage':
                             translate('finance_advice_not_giveable_because_item_not_complete',
                                       domain="PloneMeeting",
@@ -500,14 +500,14 @@ class CustomCharleroiMeetingItem(CustomMeetingItem):
                                            'proposed_to_director',):
                 # advice was already given but item was returned back to the service
                 return {'displayDefaultComplementaryMessage': False,
-                        'displayAdviceReviewState': True,
+                        'displayAdviceReviewState': False,
                         'customAdviceMessage': translate(
                             'finance_advice_suspended_because_item_sent_back_to_proposing_group',
                             domain="PloneMeeting",
                             context=item.REQUEST,
                             default="Advice is suspended because it was sent back to proposing group.")}
         res = self.context.getCustomAdviceMessageFor(advice)
-        res['displayAdviceReviewState'] = True
+        res['displayAdviceReviewState'] = False
         return res
 
     def _adviceDelayMayBeStarted(self, org_uid):