+++ CHANGES.rst
@@ -17,7 +17,7 @@ Products.MeetingCharleroi Changelog
 - Fixed `MeetingItem.mayAskAdviceAgain` to not be able to ask again a finances
   advice as `MeetingItemCharleroiCollegeWorkflowConditions.mayCorrect` was overrided.
   [gbastien]
-- Display `customAdviceMessage` and `displayAdviceReviewState` in finances advice popup.
+- Display `customAdviceMessage` in finances advice popup.
   [gbastien]
 
 4.2.2 (2025-10-29)